Hi Mr. Friedman,
Here is my code for my GPS. My main problem is that I dont know how to
label each connection with the user inputted value and I do not know
what hashmap to put as an input for my backtrace in if(v1 != null && v2
!= null). Please let me know if you could help me fix this and hopefully
finally finish this project. Thank you!

Thanks,
Kanav